Only load the platforms / image packs that're on screen
